Renamed FileUploadLink* to FileLink*
