+++ src/FaluSdk/FileLinks/FileLink.cs
@@ -1,12 +1,10 @@
 ﻿using Falu.Core;
 using System;
 
-namespace Falu.FileUploadLinks
+namespace Falu.FileLinks
 {
-    /// <summary>
-    /// A file upload link.
-    /// </summary>
-    public class FileUploadLink : FileUploadLinkPatchModel, IHasId, IHasCreated, IHasUpdated, IHasWorkspaceId, IHasLive, IHasEtag
+    /// <summary>A file link.</summary>
+    public class FileLink : FileLinkPatchModel, IHasId, IHasCreated, IHasUpdated, IHasWorkspaceId, IHasLive, IHasEtag
     {
         /// <inheritdoc/>
         public string? Id { get; set; }
@@ -17,19 +15,13 @@ namespace Falu.FileUploadLinks
         /// <inheritdoc/>
         public DateTimeOffset Updated { get; set; }
 
-        /// <summary>
-        /// Unique identifier of the file upload.
-        /// </summary>
+        /// <summary>Unique identifier of the file.</summary>
         public string? FileUploadId { get; set; }
 
-        /// <summary>
-        /// Publicly accessible URL to download the file.
-        /// </summary>
+        /// <summary>Publicly accessible URL to download the file.</summary>
         public string? Url { get; set; }
 
-        /// <summary>
-        /// Whether this link is already expired.
-        /// </summary>
+        /// <summary>Whether this link is already expired.</summary>
         public bool Expired { get; set; }
 
         /// <inheritdoc/>
