+++ src/FaluSdk/FileLinks/FileLinkPatchModel.cs
@@ -1,0 +1,16 @@
+﻿using Falu.Core;
+using System;
+using System.Collections.Generic;
+
+namespace Falu.FileLinks
+{
+    /// <summary>A model representing details that can be changed about a file link.</summary>
+    public class FileLinkPatchModel : IHasMetadata
+    {
+        /// <inheritdoc/>
+        public Dictionary<string, string>? Metadata { get; set; }
+
+        /// <summary>Time at which the link expires.</summary>
+        public DateTimeOffset? Expires { get; set; }
+    }
+}
